fastapi/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
anyio==3.3.4
    # via starlette
asgiref==3.4.1
    # via uvicorn
attrs==21.2.0
    # via jsonschema
aws-xray-sdk==2.8.0
    # via fastapi-blueprint (setup.py)
botocore==1.23.8
    # via aws-xray-sdk
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.7
    # via requests
click==8.0.3
    # via uvicorn
ecdsa==0.17.0
    # via python-jose
fastapi==0.70.0
    # via fastapi-blueprint (setup.py)
future==0.18.2
    # via aws-xray-sdk
h11==0.12.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   requests
jmespath==0.10.0
    # via botocore
jsonschema==4.2.1
    # via okdata-sdk
mangum==0.12.3
    # via fastapi-blueprint (setup.py)
okdata-aws==1.0.0
    # via fastapi-blueprint (setup.py)
okdata-sdk==0.9.1
    # via okdata-aws
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pydantic==1.8.2
    # via
    #   fastapi
    #   okdata-aws
pyjwt==2.3.0
    # via okdata-sdk
pyrsistent==0.18.0
    # via jsonschema
python-dateutil==2.8.2
    # via botocore
python-jose==3.3.0
    # via python-keycloak
python-keycloak==0.26.1
    # via okdata-sdk
requests==2.26.0
    # via
    #   okdata-aws
    #   okdata-sdk
    #   python-keycloak
rsa==4.7.2
    # via python-jose
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.2.0
    # via anyio
starlette==0.16.0
    # via
    #   fastapi
    #   okdata-aws
structlog==21.2.0
    # via okdata-aws
typing-extensions==4.0.0
    # via
    #   mangum
    #   pydantic
urllib3==1.26.7
    # via
    #   botocore
    #   requests
uvicorn==0.15.0
    # via fastapi-blueprint (setup.py)
wrapt==1.13.3
    # via aws-xray-sdk
